Preparing for Symfony 4.0 stable version (#85)
* Remove DefaultController::addFlash() method

The addFlash() method already exists in Symfony\Bundle\FrameworkBundle\Controller\ControllerTrait trait and it does exactly the same as yours. Also, if you still want to keep yours, it should be compatible with addFlash(string $type, string $message), which is supported only by PHP 7.

* Added Symfony Security Bundle

It's needed in order to secure the /admin access.

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Adding symfony/twig-bundle

* Update .travis.yml

* Remove adding Security bundle

The current version doesn't have it explicitly, it was added in as a suggestion for Symfony 4 and Flex support. Now maybe it's better to let the user decides if it's needed or not. Anyway, the Flex recipe will suggest it to the user at the end of bundle installation.